chore: adjust panel styles
